% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annotations_unique}
\alias{annotations_unique}
\title{Genome annotation data}
\format{
A data frame with 41165 rows and 8 variables:
\describe{
\item{ensembl_gene_id}{ENSEMBL Gene ID}
\item{external_gene_name}{Alternative gene name}
\item{chromosome_name}{Chromosome gene appears in}
\item{start_position}{Start position of gene in human genome}
\item{end_position}{End position of gene in human genome}
\item{strand}{Strand direction}
\item{transcript_biotype}{Biotype of transcript (i.e. protein_coding, lncRNA)}
\item{gene_biotype}{Biotype of gene (i.e. protein_coding, lncRNA)}
}
}
\usage{
annotations_unique
}
\description{
A data set containing genome annotation information of sgRNA-associated genes.
}
\keyword{datasets}
