% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{avg_expr_data}
\alias{avg_expr_data}
\title{RNA-seq expression data}
\format{
A data frame with 64258 rows and 3 variables:
\describe{
\item{gene_name}{ENSEMBL Gene ID}
\item{mean_fpkm}{Fragments Per Kilobase Million: Mean based on 9 replicates}
\item{log_fpkm}{Log transformed version of mean_fpkm (for better distribution)}
}
}
\usage{
avg_expr_data
}
\description{
A data set containing normalized RNA-seq data.
}
\keyword{datasets}
