% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{DataProcessor}
\alias{DataProcessor}
\title{The DataProcessor class}
\description{
Class for managing creating and interacting with an sqlite database of solar
analytics data.
}
\examples{

## ------------------------------------------------
## Method `DataProcessor$connect_to_new_database`
## ------------------------------------------------

dp <- DataProcessor()
dp$connect_to_new_database("database_one.db")

## ------------------------------------------------
## Method `DataProcessor$connect_to_existing_database`
## ------------------------------------------------

dp <- DataProcessor()
dp$connect_to_existing_database("database_one.db")
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-connect_to_new_database}{\code{DataProcessor$connect_to_new_database()}}
\item \href{#method-connect_to_existing_database}{\code{DataProcessor$connect_to_existing_database()}}
\item \href{#method-build_database}{\code{DataProcessor$build_database()}}
\item \href{#method-clean_data}{\code{DataProcessor$clean_data()}}
\item \href{#method-clone}{\code{DataProcessor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connect_to_new_database"></a>}}
\if{latex}{\out{\hypertarget{method-connect_to_new_database}{}}}
\subsection{Method \code{connect_to_new_database()}}{
Creates a new sqlite database with the path and name provided. All 
subsequent actions are peformed on this database unless the connection
is changed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProcessor$connect_to_new_database(db_path_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db_path_name}}{the path and name of the database to create.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{dp <- DataProcessor()
dp$connect_to_new_database("database_one.db")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connect_to_existing_database"></a>}}
\if{latex}{\out{\hypertarget{method-connect_to_existing_database}{}}}
\subsection{Method \code{connect_to_existing_database()}}{
Conects to an existing sqlite database with the path and name provided. 
All subsequent actions are peformed on this database unless the 
connection is changed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProcessor$connect_to_existing_database(db_path_name)}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{dp <- DataProcessor()
dp$connect_to_existing_database("database_one.db")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-build_database"></a>}}
\if{latex}{\out{\hypertarget{method-build_database}{}}}
\subsection{Method \code{build_database()}}{
Inserts data from csvs into the connected database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProcessor$build_database(timeseries, circuit_details, site_details)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{circuit_details}}{The path and name of a csv file containting the 
meta data of the measurements on a circuit basis.}

\item{\code{site_details}}{The path and name of a csv file containting the 
meta data of the measurements on a circuit basis.}

\item{\code{time_series}}{The path and name of a csv file containting the time
series data from solar analytics. It should contain, a column for the 
timestamp (time ending of the measurement), a column with the circuit id,
a column with the duration of the measurement, a column with the 
cumulative energy over the measurement time, and columns for volatge and
frequency measurements.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
None yet
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean_data"></a>}}
\if{latex}{\out{\hypertarget{method-clean_data}{}}}
\subsection{Method \code{clean_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProcessor$clean_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProcessor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
