% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{essential_sgrnas}
\alias{essential_sgrnas}
\title{Essential sgRNAs Dataset}
\format{
A tibble with 100 rows and 4 variables:
\describe{
\item{seq}{Sequence of the sgRNA}
\item{gene}{Gene targeted by the sgRNA}
\item{lfc}{Log fold change}
\item{fdr}{False discovery rate}
}
}
\usage{
data(essential_sgrnas)
}
\description{
A dataset containing information about essential sgRNAs.
}
\keyword{datasets}
