% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_functions.R
\name{find_essential_sgrnas}
\alias{find_essential_sgrnas}
\title{Find essential sgRNAs}
\usage{
find_essential_sgrnas(data, lfc_threshold = -1, fdr_threshold = 0.05)
}
\arguments{
\item{data}{A defined Excel sheet with sgRNA screen data using readxl}

\item{lfc_threshold}{The threshold of log fold change (LFC) to be considered essential; defaults to -1}

\item{fdr_threshold}{The threshold of False Discovery Rate (FDR) to be considered statistically significant; defaults to 0.05}
}
\value{
A dataset of sgRNAs considered essential. Includes sgRNA ID, gene name, LFC & FDR. The head of the dataset is printed.
}
\description{
Find essential sgRNAs
}
\examples{
find_essential_genes(data_file)
find_essential_genes(data_file, lfc_threshold = -2, fdr_threshold = 0.01)
}
