% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_functions.R
\name{load_if_missing}
\alias{load_if_missing}
\title{Load data files if not already loaded into environment}
\usage{
load_if_missing(object_name, file_path)
}
\arguments{
\item{object_name}{The name of the variable the data set is stored in}

\item{file_path}{The file path for the data set}
}
\value{
A message detailing whether the data was just loaded in or already existed in the environment
}
\description{
Load data files if not already loaded into environment
}
\examples{
load_if_missing("my_data_set", "data/mds.rda")
}
