% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Init.R
\name{loadTables}
\alias{loadTables}
\title{Load tables from RData files}
\usage{
loadTables(table_path, name_pattern = NA)
}
\arguments{
\item{table_path}{The path to the directory containing the RData files.}

\item{name_pattern}{An optional regular expression pattern to filter table names.}
}
\value{
A list of loaded tables.
}
\description{
This function loads tables from RData files located in the specified directory.
}
