% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ml_data}
\alias{ml_data}
\title{Data set for machine learning algorithm}
\format{
A data frame with 99389 rows and 90 variables:
\describe{
\item{base_1}{1st base in sgRNA sequence (one hot encoded): A}
\item{base_2}{1st base in sgRNA sequence (one hot encoded): C}
\item{base_3}{1st base in sgRNA sequence (one hot encoded): G}
\item{base_4}{1st base in sgRNA sequence (one hot encoded): G}
\item{base_77}{20th base in sgRNA sequence (one hot encoded): A}
}
}
\usage{
ml_data
}
\description{
A data set combining all other data sets, including positional data.
}
\keyword{datasets}
