% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Init.R
\name{readExcelFileAsTableListFromExtData}
\alias{readExcelFileAsTableListFromExtData}
\title{Read an Excel file and import each sheet as a separate data.table.}
\usage{
readExcelFileAsTableListFromExtData(simpleExcelFileName)
}
\arguments{
\item{simpleExcelFileName}{The file name of the Excel file without extension.}
}
\value{
A list of imported sheets as data.tables.
}
\description{
This function reads an Excel file specified by the 'simpleExcelFileName' parameter
from internal directory './inst/extdata/' and imports each sheet as a separate
data.table. It returns a list of imported sheets.
}
