% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Init.R
\name{readExcelFilePath}
\alias{readExcelFilePath}
\title{Reads the file path of an Excel file}
\usage{
readExcelFilePath(simpleExcelFileName)
}
\arguments{
\item{simpleExcelFileName}{The name of the Excel file without file extension.}
}
\value{
The full file path of the Excel file.
}
\description{
This function constructs the file path for an Excel file with the given file name.
}
