% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_Main.R
\name{dataprocessor}
\alias{dataprocessor}
\title{Starts the retrieval for this project. This is the main start function start the Data Processor job}
\usage{
dataprocessor()
}
\description{
Starts the retrieval for this project. This is the main start function start the Data Processor job
}
