% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dropout_data}
\alias{dropout_data}
\title{sgRNA dataset from a CRISPR screen on dropout data}
\format{
A data frame with 118461 rows and 13 variables:
\describe{
\item{sgrna}{sgRNA ID}
\item{Gene}{Gene name}
\item{GeCKO library_count}{Library count}
}
}
\usage{
dropout_data
}
\description{
sgRNA dataset from a CRISPR screen on dropout data
}
\keyword{datasets}
